requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
aom=3.6.0=h6a678d5_0
blas=1.0=mkl
blosc=1.21.3=h6a678d5_0
bottleneck=1.3.7=py39ha9d4c09_0
brotli=1.0.9=h5eee18b_8
brotli-bin=1.0.9=h5eee18b_8
brotli-python=1.0.9=py39h6a678d5_8
brunsli=0.1=h2531618_0
bzip2=1.0.8=h5eee18b_6
c-ares=1.19.1=h5eee18b_0
ca-certificates=2024.6.2=hbcca054_0
certifi=2024.6.2=pyhd8ed1ab_0
cfitsio=3.470=h5893167_7
charls=2.2.0=h2531618_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=py39h06a4308_0
cloudpickle=2.2.1=py39h06a4308_0
colorama=0.4.6=pyhd8ed1ab_0
cuda=11.6.1=0
cuda-cccl=11.6.55=hf6102b2_0
cuda-command-line-tools=11.6.2=0
cuda-compiler=11.6.2=0
cuda-cudart=11.6.55=he381448_0
cuda-cudart-dev=11.6.55=h42ad0f4_0
cuda-cuobjdump=11.6.124=h2eeebcb_0
cuda-cupti=11.6.124=h86345e5_0
cuda-cuxxfilt=11.6.124=hecbf4f6_0
cuda-driver-dev=11.6.55=0
cuda-gdb=12.5.39=0
cuda-libraries=11.6.1=0
cuda-libraries-dev=11.6.1=0
cuda-memcheck=11.8.86=0
cuda-nsight=12.5.39=0
cuda-nsight-compute=12.4.1=0
cuda-nvcc=11.6.124=hbba6d2d_0
cuda-nvdisasm=12.5.39=0
cuda-nvml-dev=11.6.55=haa9ef22_0
cuda-nvprof=12.5.39=0
cuda-nvprune=11.6.124=he22ec0a_0
cuda-nvrtc=11.6.124=h020bade_0
cuda-nvrtc-dev=11.6.124=h249d397_0
cuda-nvtx=11.6.124=h0630a44_0
cuda-nvvp=12.5.39=0
cuda-runtime=11.6.1=0
cuda-samples=11.6.101=h8efea70_0
cuda-sanitizer-api=12.5.39=0
cuda-toolkit=11.6.1=0
cuda-tools=11.6.1=0
cuda-version=12.5=3
cuda-visual-tools=11.6.1=0
cycler=0.11.0=pyhd3eb1b0_0
cyrus-sasl=2.1.28=h52b45da_1
cytoolz=0.12.2=py39h5eee18b_0
dask-core=2024.5.0=py39h06a4308_0
dav1d=1.2.1=h5eee18b_0
dbus=1.13.18=hb2f20db_0
expat=2.6.2=h6a678d5_0
ffmpeg=4.3=hf484d3e_0
fontconfig=2.14.1=h4c34cd2_2
fonttools=4.51.0=py39h5eee18b_0
freetype=2.12.1=h4a9f257_0
fsspec=2024.3.1=py39h06a4308_0
gds-tools=1.10.0.4=0
giflib=5.2.1=h5eee18b_3
glib=2.78.4=h6a678d5_0
glib-tools=2.78.4=h6a678d5_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
h5py=3.6.0=py39ha0f2276_0
hdf5=1.10.6=h3ffc7dd_1
icu=73.1=h6a678d5_0
idna=3.7=py39h06a4308_0
imagecodecs=2023.1.23=py39hc4b7b5f_0
imageio=2.19.3=pyhcf75d05_0
importlib-metadata=7.0.1=py39h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46306
jpeg=9e=h5eee18b_1
jxrlib=1.1=h7b6447c_2
kiwisolver=1.4.4=py39h6a678d5_0
krb5=1.20.1=h143b758_1
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libaec=1.0.4=he6710b0_1
libavif=0.11.1=h5eee18b_0
libbrotlicommon=1.0.9=h5eee18b_8
libbrotlidec=1.0.9=h5eee18b_8
libbrotlienc=1.0.9=h5eee18b_8
libclang=14.0.6=default_hc6dbbc7_1
libclang13=14.0.6=default_he11475f_1
libcublas=11.9.2.110=h5e84587_0
libcublas-dev=11.9.2.110=h5c901ab_0
libcufft=10.7.1.112=hf425ae0_0
libcufft-dev=10.7.1.112=ha5ce4c0_0
libcufile=1.10.0.4=0
libcufile-dev=1.10.0.4=0
libcups=2.4.2=h2d74bed_1
libcurand=10.3.6.39=0
libcurand-dev=10.3.6.39=0
libcurl=8.7.1=h251f7ec_0
libcusolver=11.3.4.124=h33c3c4e_0
libcusparse=11.7.2.124=h7538f96_0
libcusparse-dev=11.7.2.124=hbbe9722_0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20230828=h5eee18b_0
libev=4.33=h7f8727e_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libglib=2.78.4=hdc74915_0
libgomp=11.2.0=h1234567_1
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libllvm14=14.0.6=hdb19cb5_3
libnghttp2=1.57.0=h2d74bed_0
libnpp=11.6.3.124=hd2722f0_0
libnpp-dev=11.6.3.124=h3c42840_0
libnvjpeg=11.6.2.124=hd473ad6_0
libnvjpeg-dev=11.6.2.124=hb5906b9_0
libpng=1.6.39=h5eee18b_0
libpq=12.17=hdbd6064_0
libssh2=1.11.0=h251f7ec_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=h5eee18b_1
libxml2=2.10.4=hfdd30dd_2
libzopfli=1.0.3=he6710b0_0
locket=1.0.0=py39h06a4308_0
lz4-c=1.9.4=h6a678d5_1
matplotlib=3.5.1=py39h06a4308_1
matplotlib-base=3.5.1=py39ha18d171_1
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py39h5eee18b_1
mkl_fft=1.3.8=py39h5eee18b_0
mkl_random=1.2.4=py39hdb19cb5_0
mysql=5.7.24=h721c034_2
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.2.1=py39h06a4308_0
nsight-compute=2024.2.0.16=2
nspr=4.35=h6a678d5_0
nss=3.89.1=h6a678d5_0
numexpr=2.8.7=py39h85018f9_0
numpy=1.22.3=py39hf6e8229_2
numpy-base=1.22.3=py39h060ed82_2
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.0.14=h5eee18b_0
packaging=23.2=py39h06a4308_0
pandas=1.4.2=py39h295c915_0
partd=1.4.1=py39h06a4308_0
pcre2=10.42=hebb0a14_1
pillow=10.3.0=py39h5eee18b_0
pip=24.0=py39h06a4308_0
ply=3.11=py39h06a4308_0
pyparsing=3.0.9=py39h06a4308_0
pyqt=5.15.10=py39h6a678d5_0
pyqt5-sip=12.13.0=py39h5eee18b_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.19=h955ad1f_1
python-dateutil=2.9.0post0=py39h06a4308_2
python_abi=3.9=2_cp39
pytorch=1.13.1=py3.9_cuda11.6_cudnn8.3.2_0
pytorch-cuda=11.6=h867d48c_1
pytorch-mutex=1.0=cuda
pytz=2024.1=py39h06a4308_0
pywavelets=1.5.0=py39ha9d4c09_0
pyyaml=6.0.1=py39h5eee18b_0
qt-main=5.15.2=h53bd1ea_10
quaternion=2022.4.2=py39hd257fcd_0
readline=8.2=h5eee18b_0
requests=2.32.2=py39h06a4308_0
scikit-image=0.19.2=py39h51133e4_0
scipy=1.11.4=py39h5f9d8c6_0
setuptools=69.5.1=py39h06a4308_0
sip=6.7.12=py39h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.10=h6a678d5_1
sqlite=3.45.3=h5eee18b_0
tbb=2021.8.0=hdb19cb5_0
tifffile=2023.4.12=py39h06a4308_0
tk=8.6.14=h39e8969_0
tomli=2.0.1=py39h06a4308_0
toolz=0.12.0=py39h06a4308_0
torchaudio=0.13.1=py39_cu116
torchvision=0.14.1=py39_cu116
tornado=6.4.1=py39h5eee18b_0
tqdm=4.64.0=pyhd8ed1ab_0
typing_extensions=4.11.0=py39h06a4308_0
tzdata=2024a=h04d1e81_0
unicodedata2=15.1.0=py39h5eee18b_0
urllib3=2.2.1=py39h06a4308_0
wheel=0.43.0=py39h06a4308_0
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7b6447c_0
zfp=1.0.0=h6a678d5_0
zipp=3.17.0=py39h06a4308_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.5=hc292b87_2
